+++ src/app/AuthenticatedAppLayout.tsx
@@ -7,12 +7,10 @@ import { DeviceActivityReporter } from "../components/DeviceActivityReporter";
 import { UpdateProvider } from "../contexts/UpdateContext";
 import { AppShell } from "../components/layout/AppShell";
 import { ThemeProvider } from "../contexts/ThemeContext";
-import { SeoHead } from "../seo/SeoHead";
 
 export function AuthenticatedAppLayout() {
   return (
     <>
-      <SeoHead noIndex />
       <AuthProvider>
       <UserSettingsProvider>
       <ThemeProvider>